Support and expose environment vars.
Such that the program sets the mxl_project variable and environment
variables are substituted in the tmux command configuration.

This allows the launcher to execute outside of the project scope
(accepts a single argument to the directory or `package.json` file) and
run the command set whilst ensuring the directory is the project - not
the current pane path.